F_T30603 move MapValidationException to core
- adjust use statements
